Incorporate presence of $ into the DFA by adding a new accepting state.
When $ is present at the end of the regex, all the other accepting states transition to the new accepting state via `254` which would mark the end of the input (this can be a different value).
The previously accepting states become non-accepting.
Additionally in the circuit (here as example in Noir), `254` is appended to the input string to mark the end.